requirements.txt
beautifulsoup4==4.11.1
fastapi==0.89.1
mangum==0.17.0
Pillow==9.4.0
requests==2.25.1
transformers==4.25.1
uvicorn[standard]==0.20.0
